Created `variant_qc` module as RF parts were starting to clutter `basics` * Split the "pre-rf" annotation path from the RF paths. Mainly because args are quite different.
